iov_iter: add a shorten call
The generic direct write path wants to shorten its memory vector.  It
does this when it finds that it has to perform a partial write due to
LIMIT_FSIZE.  .direct_IO() always performs IO on all of the referenced
memory because it doesn't have an argument to specify the length of the
IO.

We add an iov_iter operation for this so that the generic path can ask
to shorten the memory vector without having to know what kind it is.
We're happy to shorten the kernel copy of the iovec array, but we refuse
to shorten the bio_vec array and return an error in this case.

Signed-off-by: Dave Kleikamp <dave.kleikamp@oracle.com>
Cc: Zach Brown <zab@zabbo.net>
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>

(cherry picked from commit 83c46f389f09992fe93b3fa83e68077ba002bb8d)

(cherry picked from commit 75e8299b6dd25c1423b9a59d199b3afdb198a11d)